Provide performance data and sum up resource usage of multiple processes
This fixes #2 by providing basic performance data. Since this was meant to be a quick fix the plugin now provides the performance data similar to the output data in percent. Upcoming version might change this.

This commit also fixes #1 by changing the way this plugin gets its data. Before, when there were multiple processes with the same name the first one to come up during the ps call was taken to read its resource usage and put out. Now the plugin adds all resource usage (in percent) and puts out the sum of all the processes.

Since this change was made for a customers project I commit this with my other email address.