Changed output of sum of memory usage to show memory instead of cpu
The original version of the script just used awk to get the information it needed for further processing which included the headline describing what kind of information it got from ps. The newer version that sums up the usage was stripped of this header so the header was added as plain text to keep the internal logic intact. As an error there was %CPU added to any kind of output. Since the header is stripped later on before output this change is a mere cosmetic one.

Anyway it's always better to fix such issues if one of the later changes relies on this header.